Use built-in types instead of numpy aliases
Numpy types like `np.{bool,int,float}` are just type aliases of the
built-in types. The built-in types should be preferred, as discussed
e.g. in the [numpy 1.20.0 release notes](https://numpy.org/devdocs/release/1.20.0-notes.html#using-the-aliases-of-builtin-types-like-np-int-is-deprecated)

This commit replaces all the numpy type alises with their corresponding
built-in type.

+++ dm_control/mjcf/attribute_test.py
@@ -172,7 +172,7 @@ class AttributeTest(parameterized.TestCase):
     mujoco = self._mujoco
     mujoco.optional.float_array = [3, 2, 1]
     np.testing.assert_array_equal(mujoco.optional.float_array, [3, 2, 1])
-    self.assertEqual(mujoco.optional.float_array.dtype, np.float)
+    self.assertEqual(mujoco.optional.float_array.dtype, float)
     with self.assertRaisesRegex(ValueError, 'no more than 3 entries'):
       mujoco.optional.float_array = [0, 0, 0, -10]
     with self.assertRaisesRegex(ValueError, 'one-dimensional array'):
@@ -202,7 +202,7 @@ class AttributeTest(parameterized.TestCase):
     mujoco = self._mujoco
     mujoco.optional.int_array = [2, 2]
     np.testing.assert_array_equal(mujoco.optional.int_array, [2, 2])
-    self.assertEqual(mujoco.optional.int_array.dtype, np.int)
+    self.assertEqual(mujoco.optional.int_array.dtype, int)
     with self.assertRaisesRegex(ValueError, 'no more than 2 entries'):
       mujoco.optional.int_array = [0, 0, 10]
     # failed assignment should not change the value
